Cast to float

+++ apriltag/src/main/native/cpp/AprilTagDetector.cpp
@@ -49,7 +49,7 @@ void AprilTagDetector::Results::Destroy() {
 AprilTagDetector::AprilTagDetector() : m_impl{apriltag_detector_create()} {
   // apriltag_detector_create does not appropriately initialize qtp.critical_rad
   static_cast<apriltag_detector_t*>(m_impl)->qtp.critical_rad =
-      10 * std::numbers::pi / 180;
+      static_cast<float>(10 * std::numbers::pi / 180);
 }
 
 AprilTagDetector& AprilTagDetector::operator=(AprilTagDetector&& rhs) {
